+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - transform voxco resources and variables
+- convenience function for wikidata etl pipeline
 
 ### Changes